wholesale move log parsing functions into module
